% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-left_join.R
\name{\%<<+\%}
\alias{\%<<+\%}
\title{add the external data frame to the ggsc object.}
\usage{
x \%<<+\% y
}
\arguments{
\item{x}{ggsc object, the result of \code{sc_dim}, \code{sc_feature} and \code{sc_spatial}.}

\item{y}{a data.frame, which the first column should be a barcode id, which is similar with
the x$data$\code{.BarcodeID}. It also can have another column \code{features}.}
}
\value{
ggsc object
}
\description{
add the external data frame to the ggsc object.
}
