% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalWkdeCpp}
\alias{CalWkdeCpp}
\title{Two-Dimensional Weighted Kernel Density Estimation And Mapping the Result To Original Dimension}
\usage{
CalWkdeCpp(x, w, l, h, adjust = 1, n = 400L)
}
\arguments{
\item{x}{The 2-D coordinate matrix}

\item{w}{The weighted sparse matrix, the number columns the same than the number rows than x.}

\item{l}{The limits of the rectangle covered by the grid as c(xl, xu, yl, yu)}

\item{h}{The vector of bandwidths for x and y directions, defaults to normal reference bandwidth
(see bandwidth.nrd), A scalar value will be taken to apply to both directions (see ks::hpi).}

\item{adjust}{numeric value to adjust to bandwidth, default is 1.}

\item{n}{number of grid points in the two directions, default is 400.}
}
\description{
Two-Dimensional Weighted Kernel Density Estimation And Mapping the Result To Original Dimension
}
