% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aes}
\alias{theme}
\alias{\%<+\%}
\title{Objects exported from other packages}
\value{
Depending on the re-exported function
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggfun}{\code{\link[ggfun:\%<+\%]{\%<+\%}}}

  \item{ggplot2}{\code{\link[ggplot2:aes]{aes()}}, \code{\link[ggplot2:theme]{theme()}}}
}}

