% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-dim.R
\name{sc_dim}
\alias{sc_dim}
\title{sc_dim}
\usage{
sc_dim(
  object,
  dims = c(1, 2),
  reduction = NULL,
  cells = NULL,
  slot = "data",
  mapping = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{selected dimensions (must be a two-length vector) that are used in visualization}

\item{reduction}{reduction method, default is NULL and will use the default setting store in the object}

\item{cells}{selected cells to plot (default is all cells)}

\item{slot}{slot to pull expression data from (e.g., 'count' or 'data')}

\item{mapping}{aesthetic mapping}

\item{...}{additional parameters pass to 'scattermore::geom_scattermore()'}
}
\value{
dimension reduction plot
}
\description{
sc_dim
}
\seealso{
\link[scattermore:geom_scattermore]{geom_scattermore};
}
