% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-feature.R
\name{sc_feature}
\alias{sc_feature}
\title{sc_feature}
\usage{
sc_feature(
  object,
  features,
  dims = c(1, 2),
  reduction = NULL,
  cells = NULL,
  slot = "data",
  mapping = NULL,
  ncol = 3,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{selected features (i.e., genes)}

\item{dims}{selected dimensions (must be a two-length vector) that are used in visualization}

\item{reduction}{reduction method, default is NULL and will use the default setting store in the object}

\item{cells}{selected cells to plot (default is all cells)}

\item{slot}{slot to pull expression data from (e.g., 'count' or 'data')}

\item{mapping}{aesthetic mapping}

\item{ncol}{number of facet columns if 'length(features) > 1'}

\item{...}{additional parameters pass to 'scattermore::geom_scattermore()'}
}
\value{
dimension reduction plot colored by selected features
}
\description{
sc_feature
}
