% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-spatial.R
\name{sc_spatial}
\alias{sc_spatial}
\alias{sc_spatial,Seurat-method}
\alias{sc_spatial,Seurat}
\alias{sc_spatial,SingleCellExperiment-method}
\alias{sc_spatial,SingleCellExperiment}
\title{sc_spatial}
\usage{
sc_spatial(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  ...
)

\S4method{sc_spatial}{Seurat}(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  ...
)

\S4method{sc_spatial}{SingleCellExperiment}(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = 1,
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = "v",
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{selected features to be visualized}

\item{sample.id}{the index name of sample id, which only
work with SingleCellExperiment or SpatialExperiment.}

\item{image.id}{the index name of image id, which only work
with SingleCellExperiment or SpatialExperiment.}

\item{slot}{if plotting a feature, which data will be used
(e.g., 'data', 'counts'), the assay name if object
is SingleCellExperiment or SpatialExperiment.}

\item{image.plot}{whether to display the issue image as background.}

\item{image.first.operation}{character which the first operation to
image, 'rotate' or 'mirror', default is 'rotate'.}

\item{image.rotate.degree}{integer the degree to ratate image, default is NULL.}

\item{image.mirror.axis}{character the direction to mirror the image, default is 'h'.}

\item{remove.point}{whether to remove the spot points, it is nice
if your just view the issue image, default is FALSE.}

\item{mapping}{aesthetic mapping, default is NULL.}

\item{ncol}{integer number of facet columns if 'length(features) > 1', default is 6.}

\item{...}{additional parameters.}
}
\value{
ggplot object
}
\description{
sc_spatial
}
\examples{
\donttest{
library(STexampleData)
# create ExperimentHub instance
eh <- ExperimentHub()
# query STexampleData datasets
myfiles <- query(eh, "STexampleData")
spe <- myfiles[["EH7538"]]
spe <- spe[, colData(spe)$in_tissue == 1]
set.seed(123)
genes <- rownames(spe) |> sample(6) 
p <- sc_spatial(spe, features = genes, 
                image.rotate.degree = -90, 
                image.mirror.axis = NULL, 
                ncol = 3)
}
}
