% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-spatial.R
\name{sc_spatial}
\alias{sc_spatial}
\title{sc_spatial}
\usage{
sc_spatial(object, features = NULL, slot = "data")
}
\arguments{
\item{object}{Seurat object}

\item{features}{selected features to be visualized}

\item{slot}{if plotting a feature, which data will be used (e.g., 'data', 'counts')}
}
\value{
ggplot object
}
\description{
sc_spatial
}
