% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-violin.R
\name{sc_violin}
\alias{sc_violin}
\title{sc_violin}
\usage{
sc_violin(
  object,
  features,
  cells = NULL,
  slot = "data",
  .fun = NULL,
  mapping = NULL,
  ncol = 3,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{selected features}

\item{cells}{selected cells to plot (default is all cells)}

\item{slot}{slot to pull expression data from (e.g., 'count' or 'data')}

\item{.fun}{user defined function that will be applied to selected features (default is NULL and there is no data operation)}

\item{mapping}{aesthetic mapping}

\item{ncol}{number of facet columns if 'length(features) > 1'}

\item{...}{additional parameters pass to 'ggplot2::geom_geom_violin()'}
}
\value{
violin plot to visualize feature expression distribution
}
\description{
sc_violin
}
\seealso{
\link[ggplot2:geom_violin]{geom_violin};
}
