% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-geom-annot.R
\name{sc_geom_annot}
\alias{sc_geom_annot}
\title{add the annotation layer for ggsc object}
\usage{
sc_geom_annot(
  data = NULL,
  mapping = NULL,
  pointsize = 2,
  pixels = c(512, 512),
  gap_colour = "white",
  gap_alpha = 1,
  bg_line_width = 0.3,
  gap_line_width = 0.1,
  show.legend = NA,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three
options:
If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{ggplot()}.
A \code{data.frame}, will override the plot data. the \code{data.frame}
should have a barcode id or features column.
A \code{function} will be called with a single argument, the plot
data. The return value must be a ‘data.frame’, and will be
used as the layer data. A \code{function} can be created from a
‘formula’ (e.g. ‘~ head(.x, 10)’).}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified
and \code{inherit.aes = TRUE} (the default), it is combined with the default
mapping at the top level of the plot. You must supply \code{mapping} if there is no plot mapping.}

\item{pointsize}{Radius of rasterized point. Use ‘0’ for single pixels (fastest).}

\item{pixels}{Vector with X and Y resolution of the raster, default \code{c(512,512)}.}

\item{gap_colour}{colour of gap background between the bottom background
and top point point layer, default is \code{white}.}

\item{gap_alpha}{numeric the transparency of gap background colour, default is 1.}

\item{bg_line_width}{numeric the line width of background point layer,
default is \code{0.3}.}

\item{gap_line_width}{numeric the line width of gap between the background and
top point point layer, default is \code{.1}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{layer}}.}
}
\value{
layer object
}
\description{
add the annotation layer for ggsc object
}
