% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_cell}
\alias{subset_cell}
\title{subset_cell}
\usage{
subset_cell(x, feature = "nFeature_RNA", method = "mad", n = 3)
}
\arguments{
\item{x}{Seurat object}

\item{feature}{selected features to filter out cells}

\item{method}{selected method to detect outliers to be filtered. One of 'mad', 'sd', and 'quantile'}

\item{n}{if method='quantile' and e.g. n=.98, then the lower and upper side of 1\% data will be filter out;
if method is 'mad' or 'sd' (data will be transform by \verb{log1p``), only data within the interval }median (mean) +/- n * mad (sd)` will be retained.}
}
\value{
updated object
}
\description{
subset_cell
}
