% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_feature}
\alias{subset_feature}
\title{subset_feature}
\usage{
subset_feature(x, mincell, peek = TRUE)
}
\arguments{
\item{x}{Seurat object}

\item{mincell}{feature should be detected with selected minimal cells, otherwise will be filter out}

\item{peek}{logical. If TRUE, print information of the feature and not performing subsetting}
}
\value{
updated object if peek is FALSE
}
\description{
subset_feature
}
