% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsc-package.R
\docType{package}
\name{ggsc-package}
\alias{ggsc}
\alias{ggsc-package}
\title{ggsc: Visualizing Single Cell Data}
\description{
Useful functions to visualize single cell and spatial data. It supports both 'SingleCellExperiment' and 'Seurat' objects. It also supports visualizing the data using grammar of graphics implemented in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/ggsc}
  \item Report bugs at \url{https://github.com/YuLab-SMU/ggsc/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})
}

}
\keyword{internal}
