% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsc-package.R
\docType{package}
\name{ggsc-package}
\alias{ggsc-package}
\alias{ggsc}
\title{Visualizing Single Cell Data}
\description{
The package provides useful functions to visualize single cell and spatial
data. It supports both 'SingleCellExperiment' and 'Seurat' objects.
It also supports visualizing the data using grammar of graphics
implemented in 'ggplot2'.
}
