% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-dim-utilities.R
\name{sc_dim_count}
\alias{sc_dim_count}
\title{sc_dim_count}
\usage{
sc_dim_count(sc_dim_plot)
}
\arguments{
\item{sc_dim_plot}{dimension reduction plot of single cell data}
}
\value{
a bar plot to present the cell numbers of different clusters
}
\description{
sc_dim_count
}
