% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-dim-utilities.R
\name{sc_dim_geom_ellipse}
\alias{sc_dim_geom_ellipse}
\title{sc_dim_geom_ellipse}
\usage{
sc_dim_geom_ellipse(mapping = NULL, level = 0.95, ...)
}
\arguments{
\item{mapping, }{aesthetic mapping}

\item{level}{the level at which to draw an ellipse}

\item{...}{additional parameters pass to the stat_ellipse}
}
\value{
layer of ellipse
}
\description{
sc_dim_geom_ellipse
}
\seealso{
\link[ggplot2:stat_ellipse]{stat_ellipse};
}
