% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-dim-utilities.R
\name{sc_dim_geom_label}
\alias{sc_dim_geom_label}
\title{sc_dim_geom_label}
\usage{
sc_dim_geom_label(geom = ggplot2::geom_text, ...)
}
\arguments{
\item{geom}{geometric layer (default: geom_text) to display the lables}

\item{...}{additional parameters pass to the geom}
}
\value{
layer of labels
}
\description{
sc_dim_geom_label
}
