% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-dim-utilities.R
\name{sc_dim_sub}
\alias{sc_dim_sub}
\title{sc_dim_sub}
\usage{
sc_dim_sub(subset, .column = "ident")
}
\arguments{
\item{subset}{subset of clusters to be displayed}

\item{.column}{which column represents cluster (e.g., 'ident')}
}
\value{
update plot with only subset displayed
}
\description{
sc_dim_sub
}
