% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-geom-point.R
\name{sc_geom_point}
\alias{sc_geom_point}
\title{sc_geom_point}
\usage{
sc_geom_point(mapping = NULL, ...)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{...}{additional parameters pass to 'scattermore::geom_scattermore()'}
}
\value{
layer of points
}
\description{
sc_geom_point
}
