% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-spatial.R
\name{sc_spatial}
\alias{sc_spatial}
\alias{sc_spatial,Seurat-method}
\alias{sc_spatial,Seurat}
\alias{sc_spatial,SingleCellExperiment-method}
\alias{sc_spatial,SingleCellExperiment}
\title{sc_spatial}
\usage{
sc_spatial(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  plot.pie = FALSE,
  pie.radius.scale = 0.3,
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  density = FALSE,
  grid.n = 100,
  joint = FALSE,
  joint.fun = prod,
  common.legend = TRUE,
  point.size = 5,
  ...
)

\S4method{sc_spatial}{Seurat}(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  plot.pie = FALSE,
  pie.radius.scale = 0.3,
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  density = FALSE,
  grid.n = 100,
  joint = FALSE,
  joint.fun = prod,
  common.legend = TRUE,
  point.size = 5,
  ...
)

\S4method{sc_spatial}{SingleCellExperiment}(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  plot.pie = FALSE,
  pie.radius.scale = 0.3,
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  density = FALSE,
  grid.n = 100,
  joint = FALSE,
  joint.fun = prod,
  common.legend = TRUE,
  point.size = 5,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{selected features to be visualized}

\item{sample.id}{the index name of sample id, which only
work with SingleCellExperiment or SpatialExperiment.}

\item{image.id}{the index name of image id, which only work
with SingleCellExperiment or SpatialExperiment.}

\item{slot}{if plotting a feature, which data will be used
(e.g., 'data', 'counts'), the assay name if object
is SingleCellExperiment or SpatialExperiment.}

\item{plot.pie}{logical whether plot the features with pie, default is \code{FALSE}.}

\item{pie.radius.scale}{numeric scale to the radius of pie only work with \code{plot.pie=TRUE},
default is 0.3.}

\item{image.plot}{whether to display the issue image as background.}

\item{image.first.operation}{character which the first operation to
image, 'rotate' or 'mirror', default is 'rotate'.}

\item{image.rotate.degree}{integer the degree to ratate image, default is NULL.}

\item{image.mirror.axis}{character the direction to mirror the image, default is 'h'.}

\item{remove.point}{whether to remove the spot points, it is nice
if your just view the issue image, default is FALSE.}

\item{mapping}{aesthetic mapping, default is NULL.}

\item{ncol}{integer number of facet columns if 'length(features) > 1', default is 6.}

\item{density}{whether plot the 2D weighted kernel density, default is FALSE.}

\item{grid.n}{number of grid points in the two directions to estimate 2D
weighted kernel density, default is 100.}

\item{joint}{whether joint the multiple features with \code{joint.fun},
default is FALSE.}

\item{joint.fun}{how to joint the multiple features if \code{joint = TRUE},
default is prod.}

\item{common.legend}{whether to use \code{facet_wrap} to display the multiple
\code{features}, default is TRUE.}

\item{point.size}{the size of point, default is 5.}

\item{...}{additional parameters.}
}
\value{
ggplot object
}
\description{
sc_spatial
}
\examples{
\dontrun{
library(STexampleData)
# create ExperimentHub instance
eh <- ExperimentHub()
# query STexampleData datasets
myfiles <- query(eh, "STexampleData")
spe <- myfiles[["EH7538"]]
spe <- spe[, colData(spe)$in_tissue == 1]
set.seed(123)
genes <- rownames(spe) |> sample(6) 
p <- sc_spatial(spe, features = genes, 
                image.rotate.degree = -90, 
                image.mirror.axis = NULL, 
                ncol = 3)
}
}
